ai-agents/learning-ai-agents/api-test-agent/requirements.txt
openai
pytest
requests
pyyaml
python-dotenv
